LexCPP: Also fold on square brackets
This allows to fold on array literals for languages like JavaScript
that use square brackets to declare array literals.

This should not change much for languages that only use square brackets
for array indexation as the large majority of the time the open and
close brackets are placed on the same line in these cases.